:recycle: put default behavior in subcommand "do"
